Move onUserInput to own file
