+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/0xIntuition/intuition-snap.git"
   },
   "source": {
-    "shasum": "Z4wo7QnRcYUGRZdu9SGo1QA0XQgkwQMpIKDMkThCkiE=",
+    "shasum": "E3h81ZCMg2c9Ar6zoplnIFCN4qCGusFM/0dzjic+has=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
