+++ packages/snap/src/index.tsx
@@ -34,3 +34,4 @@ export const onHomePage: OnHomePageHandler = async () => {
 };
 
 export * from './onTransaction';
+export * from './onUserInput';
